requirements.txt
django-tenants==3.6.1
django==4.2.1
psycopg2==2.9.6
djangorestframework==3.14.0
django-allauth==0.54.0
django-rest-auth-dj4==0.0.6
django-rest-passwordreset==1.3.0
drf-yasg==1.21.5
django-cors-headers==3.14.0
Pillow